Refactored code into single docker container and put credentials into .env file for security
