+++ backend/.env
@@ -1,0 +1,4 @@
+MYSQL_ROOT_PASSWORD=supersecret
+MYSQL_USER=admin
+MYSQL_PASSWORD=admin5683!
+MYSQL_DATABASE=dev
