+++ backend/backend/Dockerfile
@@ -9,8 +9,5 @@ COPY wait-for-it.sh .
 
 RUN pip install --no-cache-dir -r requirements.txt
 
-# Expose port for the Flask app
-EXPOSE 8000
-
 # Run the Flask app
 CMD ["./wait-for-it.sh", "mysql:3306", "--", "flask", "--debug", "run", "--host=0.0.0.0", "--port=8000"]
