+++ backend/docker-compose.yaml
@@ -3,24 +3,30 @@ services:
     image: mysql:latest
     container_name: dev-database
     environment:
-      MYSQL_ROOT_PASSWORD: 1234
-      MYSQL_DATABASE: dev
-      MYSQL_USER: admin
-      MYSQL_PASSWORD: admin5683!
+      MYSQL_ROOT_PASSWORD: ${MYSQL_ROOT_PASSWORD}
+      MYSQL_DATABASE: ${MYSQL_DATABASE}
+      MYSQL_USER: ${MYSQL_USER}
+      MYSQL_PASSWORD: ${MYSQL_PASSWORD}
     ports:
       - "3307:3306"
     volumes:
       - ./database/init_sql:/docker-entrypoint-initdb.d
   
-  crud:
-    build: ./crud
-    container_name: crud-service
+  backend:
+    build: ./backend
+    container_name: backend-service
+    env_file:
+      - .env
     environment:
-      - FLASK_APP=main.py
-      - DATABASE_HOST=dev-database
+        MYSQL_HOST: dev-database
+        MYSQL_PORT: 3306
+        MYSQL_USER: ${MYSQL_USER}
+        MYSQL_PASSWORD: ${MYSQL_PASSWORD}
+        MYSQL_DB: ${MYSQL_DATABASE}
+        MYSQL_POOL_SIZE: 10
     depends_on:
       - mysql
     ports:
       - "8000:8000"
     volumes:
-      - ./crud/main.py:/app/main.py
+      - ./backend/:/app/
